Move live jtouch tests out of legacy file
